refactor: create folder structure for components
