+++ src/app/projects/page.tsx
@@ -1,4 +1,4 @@
-import ProjectView from "@/components/ProjectView";
+import ProjectView from "@/components/project/project-view";
 
 
 export default function ProjectsPage() {
