--- src/components/UpdateProjectModal.module.css
@@ -1,129 +1,0 @@
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 50;
-}
-
-.modal {
-  background: white;
-  border-radius: 8px;
-  padding: 2rem;
-  width: 100%;
-  max-width: 480px;
-  position: relative;
-}
-
-.modalTitle {
-  font-size: 1.25rem;
-  font-weight: bold;
-  color: #000112;
-  margin-bottom: 1.5rem;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.formGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.label {
-  font-size: 0.875rem;
-  font-weight: 500;
-  color: #828FA3;
-}
-
-.input {
-  padding: 0.75rem;
-  border: 1px solid #E4EBFA;
-  border-radius: 4px;
-  font-size: 0.875rem;
-  width: 100%;
-  transition: border-color 0.2s;
-}
-
-.input:focus {
-  outline: none;
-  border-color: #635FC7;
-}
-
-.textarea {
-  padding: 0.75rem;
-  border: 1px solid #E4EBFA;
-  border-radius: 4px;
-  font-size: 0.875rem;
-  width: 100%;
-  min-height: 112px;
-  resize: vertical;
-  transition: border-color 0.2s;
-}
-
-.textarea:focus {
-  outline: none;
-  border-color: #635FC7;
-}
-
-.error {
-  color: #EA5555;
-  font-size: 0.875rem;
-  margin-bottom: 1rem;
-}
-
-.buttonGroup {
-  display: flex;
-  justify-content: flex-end;
-  gap: 1rem;
-  margin-top: 1.5rem;
-}
-
-.submitButton {
-  padding: 0.75rem 1.5rem;
-  background: #635FC7;
-  color: white;
-  border: none;
-  border-radius: 24px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.submitButton:hover {
-  background: #A8A4FF;
-}
-
-.submitButton:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-.cancelButton {
-  padding: 0.75rem 1.5rem;
-  background: transparent;
-  color: #635FC7;
-  border: none;
-  border-radius: 24px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: color 0.2s;
-}
-
-.cancelButton:hover {
-  color: #A8A4FF;
-}
-
-.cancelButton:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
