+++ src/components/kanbanboard/ProjectKanbanBoard.tsx
@@ -2,8 +2,8 @@
 
 import { useState, useEffect, useCallback } from 'react';
 import { getTasksByProject, updateTask } from '@/features/project/actions/task';
-import TaskModal from './TaskModal';
-import DeleteTaskModal from './DeleteTaskModal';
+import TaskModal from '../modals/task/create/TaskModal';
+import DeleteTaskModal from '../modals/task/delete/DeleteTaskModal';
 import styles from './ProjectKanbanBoard.module.css';
 import { DragDropContext, Droppable, Draggable, DropResult } from '@hello-pangea/dnd';
