+++ src/components/modals/project/create/project-modal.tsx
@@ -1,59 +1,38 @@
 'use client';
 
 import { useFormStatus } from 'react-dom';
-import { updateProject } from '@/features/project/actions/project';
+import { createProject } from '@/features/project/actions/project';
 import { useActionState } from 'react';
-import styles from './ProjectModal.module.css';
+import styles from './project-modal.module.css';
 import { createPortal } from 'react-dom';
 
-interface Project {
-  id: string;
-  title: string;
-  description?: string;
-}
-
-interface UpdateProjectModalProps {
+interface ProjectModalProps {
   isOpen: boolean;
   onClose: () => void;
-  onProjectUpdated: () => void;
-  project: Project;
+  onProjectCreated: () => void;
 }
-
-type ActionState = {
-  success: boolean;
-  project?: any;
-  error?: string;
-};
-
-const initialState: ActionState = {
-  success: false
-};
-
-const wrappedUpdateProject = async (_prevState: ActionState, formData: FormData) => {
-  const id = formData.get('id') as string;
-  return updateProject(id, Object.fromEntries(formData));
-};
 
 function SubmitButton() {
   const { pending } = useFormStatus();
   return (
     <button type="submit" className={styles.submitButton} disabled={pending}>
-      {pending ? 'Updating...' : 'Update Project'}
+      {pending ? 'Creating...' : 'Create Project'}
     </button>
   );
 }
 
-export default function UpdateProjectModal({ 
-  isOpen, 
-  onClose, 
-  onProjectUpdated, 
-  project 
-}: UpdateProjectModalProps) {
-  const [state, formAction] = useActionState(wrappedUpdateProject, initialState);
+const initialState = {
+  success: false,
+  error: null,
+  project: null
+};
 
-  // Call onProjectUpdated on success
+export default function ProjectModal({ isOpen, onClose, onProjectCreated }: ProjectModalProps) {
+  const [state, formAction] = useActionState(createProject, initialState);
+
+  // Call onProjectCreated on success
   if (state.success && isOpen) {
-    onProjectUpdated();
+    onProjectCreated();
   }
 
   if (!isOpen) return null;
@@ -61,9 +40,8 @@ export default function UpdateProjectModal({
   const modal = (
     <div className={styles.modalOverlay} onClick={onClose}>
       <div className={styles.modal} onClick={e => e.stopPropagation()}>
-        <h2 className={styles.modalTitle}>Update Project</h2>
+        <h2 className={styles.modalTitle}>Create New Project</h2>
         <form action={formAction} className={styles.form}>
-          <input type="hidden" name="id" value={project.id} />
           {state.error && !state.success && (
             <div className={styles.error}>{state.error}</div>
           )}
@@ -76,7 +54,7 @@ export default function UpdateProjectModal({
               id="title"
               name="title"
               className={styles.input}
-              defaultValue={project.title}
+              placeholder="e.g. Web Design"
               required
             />
           </div>
@@ -88,7 +66,7 @@ export default function UpdateProjectModal({
               id="description"
               name="description"
               className={styles.textarea}
-              defaultValue={project.description}
+              placeholder="e.g. Design a responsive website for a client"
             />
           </div>
           <div className={styles.buttonGroup}>
