+++ src/components/modals/project/update/update-project-modal.module.css
@@ -1,0 +1,129 @@
+.modalOverlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1000;
+}
+
+.modal {
+  background: white;
+  border-radius: 6px;
+  padding: 2rem;
+  width: 100%;
+  max-width: 480px;
+  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.1);
+}
+
+.deleteModal {
+  background: white;
+  border-radius: 6px;
+  padding: 2rem;
+  width: 100%;
+  max-width: 480px;
+  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.1);
+}
+
+.modalTitle {
+  color: #000112;
+  font-size: 1.5rem;
+  font-weight: bold;
+  margin-bottom: 1.5rem;
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
+
+.formGroup {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.label {
+  color: #828FA3;
+  font-size: 0.875rem;
+  font-weight: 500;
+}
+
+.input {
+  padding: 0.5rem 1rem;
+  border: 1px solid #E4EBFA;
+  border-radius: 4px;
+  font-size: 0.875rem;
+  color: #000112;
+}
+
+.textarea {
+  padding: 0.5rem 1rem;
+  border: 1px solid #E4EBFA;
+  border-radius: 4px;
+  font-size: 0.875rem;
+  color: #000112;
+  min-height: 100px;
+  resize: vertical;
+}
+
+.buttonGroup {
+  display: flex;
+  gap: 1rem;
+  margin-top: 1rem;
+}
+
+.cancelButton {
+  background: rgba(99, 95, 199, 0.1);
+  color: #635FC7;
+  border: none;
+  padding: 0.75rem 1.5rem;
+  border-radius: 24px;
+  font-weight: 500;
+  cursor: pointer;
+  transition: background-color 0.2s;
+  flex: 1;
+}
+
+.cancelButton:hover {
+  background: rgba(99, 95, 199, 0.25);
+}
+
+.deleteConfirmButton {
+  background-color: #EA5555;
+  color: white;
+  border: none;
+  padding: 0.75rem 1.5rem;
+  border-radius: 24px;
+  font-weight: 500;
+  cursor: pointer;
+  transition: background-color 0.2s;
+  flex: 1;
+}
+
+.deleteConfirmButton:hover {
+  background-color: #FF9898;
+}
+
+.deleteConfirmButton:disabled {
+  background-color: #FF9898;
+  cursor: not-allowed;
+}
+
+.error {
+  color: #EA5555;
+  font-size: 0.875rem;
+  margin-bottom: 1rem;
+}
+
+.deleteMessage {
+  color: #828FA3;
+  font-size: 0.875rem;
+  line-height: 1.5;
+  margin-bottom: 1.5rem;
+}
