+++ src/components/sidebar/sidebar.tsx
@@ -3,10 +3,10 @@
 import { useState, useEffect, useCallback } from 'react';
 import Link from 'next/link';
 import { useSession, signOut } from 'next-auth/react';
-import styles from './Sidebar.module.css';
-import ProjectModal from './ProjectModal';
-import UpdateProjectModal from './UpdateProjectModal';
-import DeleteProjectModal from './DeleteProjectModal';
+import styles from './sidebar.module.css';
+import ProjectModal from '../modals/project/create/project-modal';
+import UpdateProjectModal from '../modals/project/update/update-project-modal';
+import DeleteProjectModal from '../modals/project/delete/delete-project-modal';
 
 interface Project {
   id: string;
